feat(radar_lidar): redesign offline registration to escape local minima
## Background
Offline registration (offline_test_node) produced inaccurate poses and got
trapped in local minima. Root cause: the initial pose dropped z (the ~4 m
sensor height) and pitch. The old init_pose hard-coded z=0 with yaw only, so
GICP started from a ground-level, level-gaze pose and fell into the wrong
basin. This was not an iteration-count or threshold problem.

## Core redesign (offline_test_node)
- look-at 6DOF init: derive yaw+pitch from sensor position toward a look-at
  target, restoring the missing z/pitch
- Local multi-start yaw search: grid search around the look-at yaw (not a
  blind 360 sweep), using the available time budget
- Normalized KdTree scoring: compute inlier ratio + inlier RMSE against the
  map pcl_tree so candidates are comparable, replacing the non-comparable
  small_gicp result.error
- Fine registration: converge from the best candidate, accept only if better
- pose.json output: includes frame annotation, inlier ratio, and RMSE

## Coordinate-frame normalization
- Lock the working frame to center-origin / meters / Z-up. When emitting to
  the referee system, convert via (x+14)*100 / (y+7.5)*100 to the official
  2026 protocol V1.3.0 frame (bottom-left origin, cm, 2D)
- Generate the normalized map directly from FBX via tools/model_to_map --y-up,
  removing the per-run /tmp conversion hack (map_y_up demoted to a deprecated
  compatibility fallback)

## GICP robustness (config/localization)
- Relax convergence thresholds: rotation_eps 0.001745 -> 0.03,
  translation_eps 0.001 -> 0.1
- On non-convergence, fall back to the best transform and propagate the
  converged flag correctly

## Modern C++ refactor
- spherical_grid / frame_accumulator / dynamic_cloud: merge point clouds via
  ranges views
- Unify float -> double (cluster / dynamic_cloud configs)
- M_PI -> std::numbers::pi

## Tooling & repo cleanup
- Relocate logger to tools/logger/ and remove the single-purpose radar_common
  package
- Move make_synth_scan.py to top-level tools/, add tools/README.md
- Delete one-shot normalize_map (superseded by model_to_map --y-up)
- Add offline_visualization.hpp + tests; add offline_registration launch/config
- Merge TODO into docs/todo.md; mount host SSH key + gh auth in devcontainer

+++ ros_ws/src/radar_lidar/include/radar_lidar/cluster.hpp
@@ -11,9 +11,9 @@
 namespace radar {
 
 struct ClusterConfig {
-    float cluster_tolerance = 0.25f; // 欧氏聚类距离阈值
-    int min_cluster_size    = 5;
-    int max_cluster_size    = 1000;
+    double cluster_tolerance = 0.25;
+    int min_cluster_size     = 5;
+    int max_cluster_size     = 1000;
 };
 
 struct ClusterResult {
